refactor(trusty-code): split llm/types.rs into message/request/response/usage submodules (500-line cap, PR #1040)
types.rs was 633 lines and blocked the line-cap CI gate. Split into four
focused sibling modules under crates/trusty-code/src/llm/:
- message.rs  (172 lines) — ChatMessage + role constructors
- request.rs  (252 lines) — ToolCall/FunctionCall, ToolDefinition/FunctionDefinition, ChatRequest
- response.rs (240 lines) — AssistantMessage, ChatChoice, ChatResponse + helpers
- usage.rs    (106 lines) — UsageBlock + into_token_usage mapping

mod.rs re-exports the same public surface so all downstream imports are
unchanged. All 203 tests pass; .line-cap-allowlist.tsv is not extended.
